C-style block decommenting
This was a nightmare with the current line-by-line processing that the decommenter uses. Using it on lines where there's code on the line with a start or end symbol will cause abnormal behavior. Later, the code can be restructured to do multi-line regex which will massively simplify the code, but for now this works.