Recommenting "works"
Recommenting technically works, in that it runs without errors and all of the comments are in the document. However, there are spacing issues and the program treats lines of block comments as newline comments and adds //s for c-style which isn't desired. And it still has issues with code and comments on the same line. This will get pushed to dev, but it will need to be revamped in the near future.